Move listeners to their respective module, and removed more unneeded stuff
