+++ src/main/java/com/strangeone101/holoitemsapi/customdamage/CIDamageEntityEvent.java
@@ -1,4 +1,4 @@
-package com.strangeone101.holoitemsapi.event;
+package com.strangeone101.holoitemsapi.customdamage;
 
 import com.strangeone101.holoitemsapi.util.CustomDamageSource;
 import org.bukkit.entity.Entity;
