+++ src/main/java/com/strangeone101/holoitemsapi/recipe/CraftListener.java
@@ -1,16 +1,17 @@
-package com.strangeone101.holoitemsapi.listener;
+package com.strangeone101.holoitemsapi.recipe;
 
 import com.strangeone101.holoitemsapi.CustomItemManager;
-import com.strangeone101.holoitemsapi.recipe.NonConsumableChoice;
-import com.strangeone101.holoitemsapi.recipe.RecipeBuilder;
-import com.strangeone101.holoitemsapi.recipe.RecipeManager;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.inventory.CraftItemEvent;
 import org.bukkit.event.inventory.PrepareItemCraftEvent;
 import org.bukkit.event.player.PlayerRecipeDiscoverEvent;
-import org.bukkit.inventory.*;
+import org.bukkit.inventory.CraftingInventory;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.RecipeChoice;
+import org.bukkit.inventory.ShapedRecipe;
+import org.bukkit.inventory.ShapelessRecipe;
 import org.bukkit.scheduler.BukkitRunnable;
 import xyz.holocons.mc.holoitemsrevamp.HoloItemsRevamp;
