--- src/main/java/com/strangeone101/holoitemsapi/util/ArmorStand3DItem.java
@@ -1,12 +1,0 @@
-package com.strangeone101.holoitemsapi.util;
-
-import org.bukkit.entity.ArmorStand;
-import org.bukkit.entity.Entity;
-import org.bukkit.inventory.ItemStack;
-
-public abstract class ArmorStand3DItem {
-
-    private Entity owner;
-    private ArmorStand armorStand;
-    private ItemStack item;
-}
