+++ src/main/java/com/strangeone101/holoitemsapi/util/CustomDamage.java
@@ -1,7 +1,7 @@
 package com.strangeone101.holoitemsapi.util;
 
-import com.strangeone101.holoitemsapi.event.CIDamageEntityEvent;
-import com.strangeone101.holoitemsapi.event.CIEntityDamageEntityEvent;
+import com.strangeone101.holoitemsapi.customdamage.CIDamageEntityEvent;
+import com.strangeone101.holoitemsapi.customdamage.CIEntityDamageEntityEvent;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
